Created Model/Episode.swift. Removed the stuct Episode from the EpisodesController.swift and transferred it to Episode.swift.
