+++ Podcast/Podcast/Models/Episode.swift
@@ -1,0 +1,22 @@
+//
+//  Episode.swift
+//  Podcast
+//
+//  Created by Jaycell on 07/09/2019.
+//  Copyright © 2019 Jaycell. All rights reserved.
+//
+
+import Foundation
+import FeedKit
+
+struct Episode {
+    let title: String
+    let pubDate: Date
+    let description: String
+    
+    init(feedItem: RSSFeedItem) {
+        self.title = feedItem.title ?? ""
+        self.pubDate = feedItem.pubDate ?? Date()
+        self.description = feedItem.description ?? ""
+    }
+}
